Reviewed
Instead of using .then() function on promises, we can use async/await keywords.
To render something when a route is matched, we can also pass it as a child instead of using the render props
On the search page, if a book is not assigned to a shelf, the option "None" should be selected in the dropdown menu.
Dynamics Shelves,